# Remove explicit obj.isNull() calls and let core return null values instead