@RedBeard0531 2023-03-29
> I'm left wondering if we could make it even faster if the Core code didn't do the null check on non-optional properties. As in a variant of getAny that would expect the value to never be null.

Probably not. Most of the expense here is because we were adding an extra JS->C++ call. Even if it did nothing, that still is moderately expensive. Not very expensive, but everything else we are doing is so cheap that even small costs can matter.

Additionally, the embedded "null check" that core does as part of `getAny` (and `getLinkedObject`) is cheaper than a separate call to `isNull`. The embedded null check is just an extra branch after performing the lookup which is essentially free. On the other hand, calling `isNull` basically needs to do its own separate lookup just to check for nullness, then `getAny` the same lookup again to fetch the value.